🔨 유저 프로필 리팩터링
