+++ hooks/useExpandedState.ts
@@ -1,0 +1,12 @@
+import { useState } from 'react';
+
+export const useExpandedState = () => {
+  const [isExpanded, setIsExpanded] = useState(false);
+
+  const toggleExpand = () => setIsExpanded(!isExpanded);
+
+  return {
+    isExpanded,
+    toggleExpand,
+  };
+};
